Lots of changes and debugging
CameraComponent tested and working
set CameraComponent as active view each update
SpriteRenderer up and working, needs to interface with RsourceManager
ResourceManager up, needs more testing
changed ID type from int to uint64
switched from tabs to spaces

+++ MyGameEngine/AEntity.h
@@ -4,30 +4,30 @@
 #include "AComponent.h"
 
 namespace MGE {
-	class AEntity : public sf::Transformable, public sf::Drawable {
-	protected:
-		unsigned int _ID;
-		std::string _name;
-		std::vector<AComponent*> _components;
+    class AEntity : public sf::Transformable, public sf::Drawable {
+    protected:
+        uint64_t _ID;
+        std::string _name;
+        std::vector<AComponent*> _components;
 
-	public:
-		AEntity();
-		AEntity(std::string name);
-		AEntity(std::string name, float x, float y);
+    public:
+        AEntity();
+        AEntity(std::string name);
+        AEntity(std::string name, float x, float y);
 
-		virtual void Awake();
-		virtual void Start();
-		virtual void Update(float deltaTime);
-		virtual void LateUpdate(float deltaTime);
+        virtual void Awake();
+        virtual void Start();
+        virtual void Update(float deltaTime);
+        virtual void LateUpdate(float deltaTime);
 
-		const std::vector<AComponent*> getComponents();
-		void attachComponent(AComponent* component);
+        const std::vector<AComponent*> getComponents();
+        void attachComponent(AComponent* component);
 
-		unsigned int getID();
+        uint64_t getID();
 
-		void draw(sf::RenderTarget& target, sf::RenderStates states) const;
+        void draw(sf::RenderTarget& target, sf::RenderStates states) const;
 
-		std::string E();
-	};
+        std::string E();
+    };
 }
